commenter/requirements.txt
asgiref==3.5.2
asttokens==2.2.0
backcall==0.2.0
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
Django==4.1.3
django-crispy-forms==1.14.0
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.1.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==24.0.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.6.0
tzdata==2022.6
wcwidth==0.2.5
webencodings==0.5.1
